Change discord.Intents.all() to discord.Intents.default()
In this update, we've modified the way discord client handles the intents. Instead of using discord.Intents.all(), we are now using discord.Intents.default() which gives us only the ticker updates for name, icon, type and owner changes. We have also explicitly allowed 'members' and 'messages' intents, as they are necessary for our program functionality. These changes are expected to reduce the amount of data we process and thus improve performance.